Light refactoring to remove code duplication.
