+++ src/lib.rs
@@ -7,3 +7,4 @@ pub mod ecdsa;
 pub mod rsa;
 pub mod tests;
 pub mod syscon;
+pub mod controller;
